pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "makerworks-backend"
version = "0.1.0"
description = "FastAPI backend for MakerWorks platform"
authors = [{ name = "Stephen", email = "stephen@example.com" }]
license = "MIT"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
  "fastapi>=0.111.0",
  "uvicorn>=0.30.1",
  "loguru>=0.7.2",
  "sqlalchemy>=2.0.30",
  "asyncpg>=0.29.0",
  "alembic>=1.13",
  "python-dotenv>=1.0.1",
  "python-magic>=0.4.27",
  "pydantic[email]>=2.7.1",
  "pydantic-settings>=2.2.1",  # ✅ Required for BaseSettings in Pydantic v2
  "redis>=5.0.4",
  "celery>=5.3.6",
  "stripe>=9.5.0",
  "psutil>=5.9.8",
  "prometheus_client>=0.20.0",
  "greenlet>=3.2.3,<4.0.0",
  "psycopg2-binary>=2.9.10,<3.0.0",
  "gunicorn>=21.2.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "python-multipart>=0.0.9",
  "httpx>=0.27.0",
  "pynvml>=11.5.0",
  "jwcrypto>=1.5.4",
  "pillow>=10.3.0"
]

[project.optional-dependencies]
dev = [
  "black>=24.4.2",
  "ruff>=0.4.4",
  "pytest>=8.2.2",
  "httpx>=0.27.0"
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"
fix = true
exclude = [".git", ".mypy_cache", ".ruff_cache", ".venv", "build", "dist"]
select = ["E", "F", "B", "I", "UP", "N", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.isort]
known-first-party = ["app"]
combine-as-imports = true
force-single-line = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"